Re-enable unchecked benchmark now enabled by #242
